requirements.txt
python-dotenv
langchain-google-genai
langchain
langchain_community
tavily-python